WIP: OMAP4: add coresight components
The base address are from Table 2-2 L3_EMU Memory Space Mapping of the
OMAP4430 TRM and the connections are from Figure 28-2 Processor Trace
Flow of the same document.
https://focus.ti.com/pdfs/wtbu/OMAP4430_ES2.0_Public_TRM_vJ.pdf

This is also a very useful reference:
http://www2.lauterbach.com/pdf/arm_app_coresight.pdf

=== HWMOD Hacks

The hacks to enable all HWMODs are to enable the required
clocks.  The clocks required are as described here:

http://permalink.gmane.org/gmane.linux.ports.arm.kernel/112477

and equivalent to these register writes:

The presence of the PERIPHID can be used to check if the correct clocks
have been enabled (the drivers won't probe otherwise):

0x00000090

=== Usage

Read "How to use" in Documentation/trace/coresight.txt.  This works:

=== Decoding

These tools decode something:

./ptm2human -i ~/trace.bin -c 4
https://github.com/hwangcc23/ptm2human

python etm-objdump.py trace.bin --trace-decoder-options="--pft-1.1"
https://lists.linaro.org/pipermail/linaro-dev/2012-November/014439.html