kasan: allow arch overide of number of ptes
ARM stores the Linux PTEs and the Hardware PTEs after each other in the
second level of the page table, as described in
arch/arm/include/asm/pgtable-2level.h.  So ARM needs kasan_zero_ptes to
double its size in order to avoid overwriting memory in set_pte_at().

To this end, make the number of elements in kasan_zero_ptes overridable
by the architecture.

Signed-off-by: Rabin Vincent <rabin@rab.in>