sockmap: fix splice
If a socket is added to a sockmap with a verdict program which returns
SK_PASS, splice(2) does not work on the socket.

The stream parser/verdict code removes skbs from the receive queue,
checks them using the bpf program, and then re-queues them onto a
separate queue (psock->ingress_msg).  The sockmap code modifies the TCP
recvmsg hook to check this second queue also so that works. But the
splice_read hooks is not modified and the deafult tcp_read_splice() only
reads the normal receive queue so it never sees the skbs which have been
re-queued.

Fix it by using copy_splice_read() when replacing the proto for the
sockmap.  This should be replaced with a more efficient custom version.

Signed-off-by: Vincent Whitchurch <vincent.whitchurch@datadoghq.com>